Rename grouped_slurm->slurm and slurm->legacy_slurm in tests
