+++ tests/test_full_workflow.py
@@ -50,9 +50,9 @@ async def test_full_workflow(
     override_settings_factory,
 ):
 
-    if backend == "slurm" or backend == "local":
+    if backend == "legacy_slurm" or backend == "local":
         FRACTAL_SLURM_CONFIG_FILE = testdata_path / "old_slurm_config.json"
-    elif backend == "grouped_slurm":
+    elif backend == "slurm":
         FRACTAL_SLURM_CONFIG_FILE = testdata_path / "slurm_config.json"
     override_settings_factory(
         FRACTAL_RUNNER_BACKEND=backend,
@@ -62,7 +62,7 @@ async def test_full_workflow(
 
     debug(f"Testing with {backend=}")
 
-    if backend in ["slurm", "grouped_slurm"]:
+    if backend in ["slurm", "legacy_slurm"]:
         request.getfixturevalue("monkey_slurm")
         request.getfixturevalue("relink_python_interpreter")
         request.getfixturevalue("cfut_jobs_finished")
@@ -224,9 +224,9 @@ async def test_failing_workflow_TaskExecutionError(
     override_settings_factory,
 ):
 
-    if backend == "slurm" or backend == "local":
+    if backend == "legacy_slurm" or backend == "local":
         FRACTAL_SLURM_CONFIG_FILE = testdata_path / "old_slurm_config.json"
-    elif backend == "grouped_slurm":
+    elif backend == "slurm":
         FRACTAL_SLURM_CONFIG_FILE = testdata_path / "slurm_config.json"
     override_settings_factory(
         FRACTAL_RUNNER_BACKEND=backend,
@@ -236,7 +236,7 @@ async def test_failing_workflow_TaskExecutionError(
     )
 
     debug(f"Testing with {backend=}")
-    if backend in ["slurm", "grouped_slurm"]:
+    if backend in ["slurm", "legacy_slurm"]:
         request.getfixturevalue("monkey_slurm")
         request.getfixturevalue("relink_python_interpreter")
         request.getfixturevalue("cfut_jobs_finished")
@@ -361,7 +361,7 @@ def _auxiliary_run(slurm_user, sleep_time):
     loop.close()
 
 
-@pytest.mark.parametrize("backend", ["slurm", "grouped_slurm"])
+@pytest.mark.parametrize("backend", ["slurm", "legacy_slurm"])
 @pytest.mark.slow
 async def test_failing_workflow_JobExecutionError(
     backend,
@@ -380,9 +380,9 @@ async def test_failing_workflow_JobExecutionError(
     cfut_jobs_finished,
 ):
 
-    if backend == "slurm" or backend == "local":
+    if backend == "legacy_slurm" or backend == "local":
         FRACTAL_SLURM_CONFIG_FILE = testdata_path / "old_slurm_config.json"
-    elif backend == "grouped_slurm":
+    elif backend == "slurm":
         FRACTAL_SLURM_CONFIG_FILE = testdata_path / "slurm_config.json"
     override_settings_factory(
         FRACTAL_RUNNER_BACKEND=backend,
