+++ tests/test_unit_slurm_batching_heuristics.py
@@ -4,7 +4,7 @@ from pathlib import Path
 import pytest
 from devtools import debug
 
-from fractal_server.app.runner._grouped_slurm._batching_heuristics import (
+from fractal_server.app.runner._slurm._batching_heuristics import (
     heuristics,
 )
