+++ tests/test_unit_slurm_config.py
@@ -6,10 +6,10 @@ from devtools import debug
 
 from .fixtures_tasks import MockTask
 from .fixtures_tasks import MockWorkflowTask
-from fractal_server.app.runner._grouped_slurm._slurm_config import (
+from fractal_server.app.runner._slurm._slurm_config import (
     get_default_slurm_config,
 )
-from fractal_server.app.runner._grouped_slurm._slurm_config import (
+from fractal_server.app.runner._slurm._slurm_config import (
     get_slurm_config,
 )
